+++ airbyte-workers/src/test/java/io/airbyte/workers/SingerSyncWorkerTest.java
@@ -22,23 +22,26 @@
  * SOFTWARE.
  */
 
-package io.airbyte.workers.protocols.singer;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+package io.airbyte.workers;
 
 import io.airbyte.config.StandardSync;
 import io.airbyte.config.StandardSyncInput;
 import io.airbyte.config.StandardTapConfig;
 import io.airbyte.config.StandardTargetConfig;
 import io.airbyte.singer.SingerMessage;
-import io.airbyte.workers.SingerSyncWorker;
-import io.airbyte.workers.TestConfigHelpers;
+import io.airbyte.workers.protocols.singer.DefaultSingerTap;
+import io.airbyte.workers.protocols.singer.DefaultSingerTarget;
+import io.airbyte.workers.protocols.singer.SingerMessageUtils;
+import io.airbyte.workers.protocols.singer.SingerTap;
+import io.airbyte.workers.protocols.singer.SingerTarget;
 import java.nio.file.Path;
 import java.util.Optional;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.junit.jupiter.api.Test;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 class SingerSyncWorkerTest {
